Updates (see des)
Optimized some of interpreter (more needs to be done for sure)
Added the ability to draw circles
Added another side panel that is accessible by clicking display code on the top right hand corner. It will display the code that is translated from the user inputs, along with line numbers next to them.
Fixed the issue of the text box not scrolling to the bottom after having more than 35 or so lines.

To be implemented next:
The turtle cannot move outside the given canvas. Instead, it will trace the border.
Code display was implemented to begin work on repeating commands, which will also be implemented next