Update (see description)
The turtle now "traces the border" when moving too far in any direction. For example, try:

turn left 30
forward 400

Bugs to fix:

This feature does not work if the turtle goes too far vertically and horizontally
The turtle does not move correctly if it is already at the border and must move past another border (i.e. forward 400, backwards 700)